skeleton/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
absl-py==0.9.0            # via tensorboard, tensorflow, tensorflow-estimator
astor==0.8.1              # via tensorflow
click==7.1.2              # via flask
flask==1.1.2              # via skeleton (setup.py)
gast==0.2.2               # via tensorflow
grpcio==1.28.1            # via tensorboard, tensorflow
h5py==2.10.0              # via keras-applications
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
markdown==3.2.1           # via tensorboard
markupsafe==1.1.1         # via jinja2
mock==3.0.5               # via tensorflow-estimator
numpy==1.18.4             # via h5py, keras-applications, keras-preprocessing, opencv-contrib-python, skeleton (setup.py), tensorboard, tensorflow, tensorflow-estimator
opencv-contrib-python==3.4.9.33  # via skeleton (setup.py)
protobuf==3.11.3          # via tensorboard, tensorflow
pyyaml==5.3.1             # via skeleton (setup.py)
six==1.14.0               # via absl-py, grpcio, h5py, keras-preprocessing, mock, protobuf, tensorboard, tensorflow, tensorflow-estimator
tensorboard==1.13.1       # via tensorflow
tensorflow-estimator==1.13.0  # via tensorflow
tensorflow==1.13.2        # via skeleton (setup.py)
termcolor==1.1.0          # via tensorflow
werkzeug==1.0.1           # via flask, tensorboard
wheel==0.34.2             # via tensorboard, tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
